+++ pyaud/src/modules.py
@@ -11,9 +11,11 @@ from . import (
     DeployDocs,
     EnterDir,
     Git,
+    HashCap,
     LineSwitch,
     PyaudSubprocessError,
     Subprocess,
+    Tally,
     check_command,
     colors,
     config,
@@ -32,6 +34,7 @@ def make_audit(**kwargs: Union[bool, str]) -> int:
     """
     audit_modules: List[Callable[..., Any]] = [
         make_format,
+        make_imports,
         make_typecheck,
         make_unused,
         make_lint,
@@ -427,3 +430,36 @@ def make_whitelist(**kwargs: Union[bool, str]) -> int:
             )
 
     return 0
+
+
+def make_imports():
+    """Sort imports with ``isort``. ``Black`` and ``isort`` clash in
+    some areas when it comes to ``Black`` and sorting imports. To
+    avoid running into false positives when running both in conjunction
+    run ``Black`` straight after. Use ``HashCap`` to determine if any
+    files have changed for presenting data to user.
+    """
+    changed = []
+    isort = Subprocess("isort")
+    black = Subprocess("black")
+    for item in pyitems.files:
+        if os.path.isfile(item):
+            with HashCap(item) as cap:
+                isort.call(item, capture=True)
+                black.call(item, devnull=True)
+
+            if not cap.compare:
+                changed.append(
+                    os.path.relpath(item, environ.env["PROJECT_DIR"])
+                )
+                print(isort.stdout.strip())
+
+    if changed:
+        command = "isort {}".format(" ".join(changed))
+        raise PyaudSubprocessError(returncode=1, cmd=command)
+
+    colors.green.bold.print(
+        "Success: no issues found in {} source files".format(
+            Tally.pyfiles(*pyitems.items)
+        )
+    )